ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.7
python-multipart==0.0.6
pdfminer.six==20221105
spacy==3.7.2
httpx==0.25.2
pydantic==2.5.0
langdetect==1.0.9
deep-translator==1.11.4
